wip vanishing route line working

+++ librouteline/src/main/java/com/mapbox/navigation/ui/internal/route/RouteConstants.java
@@ -21,7 +21,7 @@ public class RouteConstants {
   public static final String ARROW_HEAD_ICON_CASING = "mapbox-navigation-arrow-head-icon-casing";
   public static final int MAX_DEGREES = 360;
   public static final String ARROW_HEAD_CASING_LAYER_ID = "mapbox-navigation-arrow-head-casing-layer";
-  public static final Float[] ARROW_HEAD_CASING_OFFSET = {0f, -7f};
+  public static final Double[] ARROW_HEAD_CASING_OFFSET = {0.0, -7.0};
   public static final String ARROW_HEAD_LAYER_ID = "mapbox-navigation-arrow-head-layer";
   public static final Float[] ARROW_HEAD_OFFSET = {0f, -7f};
   public static final int MIN_ARROW_ZOOM = 10;
